Fixed compilation warnings of C code.

+++ src/cfuncs/src/dwt.c
@@ -75,8 +75,8 @@ static PyObject *wlikelihood(PyObject *self, PyObject *args){
   sigmaw = INDd(params, 2);
 
   /* Get data array size:                                             */
-  rsize = PyArray_DIM(res, 0);   /* Get residuals vector size         */
-  M = ceil(1.0*log2(rsize));     /* Number of scales                  */
+  rsize = (int)PyArray_DIM(res, 0);   /* Get residuals vector size    */
+  M = ceil(1.0*log2(rsize));          /* Number of scales             */
 
   /* Expand res to a size proportional to 2^M (zero padding)          */
   wrsize = (int)pow(2, M);     /* Expanded size                       */
@@ -160,7 +160,7 @@ static PyObject *daubechies4(PyObject *self, PyObject *args){
     return NULL;
 
   /* Get size of input vector:                                      */
-  vsize = PyArray_DIM(vector, 0);
+  vsize = (int)PyArray_DIM(vector, 0);
 
   /* Allocate memory for pointer with the data:                     */
   ptrvector = malloc(vsize * sizeof(double));
